Add slight optimizations & modify randomizer's call-to-action message
Overhead from NumPy function calls is much larger than the savings in terms of memory space and time. This is because we use non-vectorized operations in the randomized algorithm. Hence, we remove the usage of NumPy and stick with native Python lists. Method to find integer ceiling division has also been optimized.