Optimize group randomizer functions
This commit adds several slight optimizations for the group randomizer:

1. Generate partitions instead of compositions since it requires less
memory (and hence, faster) and since we shuffle the selected
configuration anyway. This is because for a given non-negative integer
n, the partition function `p(n)` is always less than or equal to the
number of compositions of n (1 for 0, 2^(n-1) for n>=1).

2. For integer partitions, the sigma restriction function would impose
the parts to be monotonically increasing (non-decreasing). Hence, we can
remove the usage of `min()` and `max()` (which have O(n) time
complexity), and instead, we can simply only check the first and last
parts (with O(1) time complexity).

3. Set `m = settings.MIN_GROUP_SIZE` since this would prevent the
generation of integer partitions where the first part is less than the
minimum value, allowing us to use even less memory space.

The value of `settings.RANDOMIZER_CHANNEL_SIZE_THRESHOLD` has also been
increased accordingly since we can now afford to properly randomize
channels of larger sizes.

Signed-off-by: James Raphael Tiovalen <jamestiotio@gmail.com>